# src/main/resources/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Movie Ticket Booking API
  description: 'Browse screenings, reserve seats, complete bookings, cancel bookings, and administer the
    local catalog. This document is the source of truth: request/response models and the controller interfaces
    are generated from it at build time.'
  version: v1
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
...
  /api/v1/seat-reservations:
    post:
      tags:
      - Customer reservations
      summary: Reserve one to ten screening seats atomically
      operationId: createReservation
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        $ref: '#/components/parameters/IdempotencyKey'
        schema:
          type: string
          maxLength: 120
          minLength: 0
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
            example:
              screeningSeatIds:
              - 00000000-0000-4000-8000-000000000006
              - 00000000-0000-4000-8000-000000000007
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /admin/api/v1/screenings:
    get:
      tags:
      - Admin screenings
      operationId: listScreenings
      parameters:
      - name: auditoriumId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: movieId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseScreeningResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
      - basicAuth: []
    post:
      tags:
      - Admin screenings
      operationId: createScreening
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScreeningRequest'
            example:
              auditoriumId: 00000000-0000-4000-8000-000000000003
              movieId: 00000000-0000-4000-8000-000000000002
              pricingPlanId: 00000000-0000-4000-8000-000000000004
              startTime: '2030-01-07T12:30:00Z'
              endTime: '2030-01-07T14:30:00Z'
              refundPolicyId: 00000000-0000-4000-8000-000000000005
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreeningResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /api/v1/seat-reservations/{id}:
    get:
      tags:
      - Customer reservations
      summary: Get an owned seat reservation
      operationId: getReservation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    delete:
      tags:
      - Customer reservations
      summary: Release an owned seat reservation idempotently
      operationId: releaseReservation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
  /api/v1/screenings:
    get:
      tags:
      - Public screenings
      summary: Browse screenings with an opaque start-time cursor
      operationId: browseScreenings
      parameters:
      - name: cityId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      - name: movieId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: theaterId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/CursorLimit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPageResponseScreeningSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/screenings/{id}:
    get:
      tags:
      - Public screenings
      summary: Get screening and price details
      operationId: browseScreening
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreeningDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/screenings/{id}/seats:
    get:
      tags:
      - Public screenings
      summary: Get the effective seat availability for a screening
      operationId: browseScreeningSeats
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeatAvailability'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/api/v1/screenings/{id}:
    get:
      tags:
      - Admin screenings
      operationId: getScreening
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreeningResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    delete:
      tags:
      - Admin screenings
      operationId: cancelScreening
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
  /admin/api/v1/screenings/{id}/prices:
    get:
      tags:
      - Admin screenings
      operationId: listScreeningPrices
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScreeningPriceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
components:
  schemas:
...
    CreateReservationRequest:
      type: object
      properties:
        screeningSeatIds:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 10
          minItems: 1
      required:
      - screeningSeatIds
    ReservationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        screeningId:
          type: string
          format: uuid
        screeningSeatIds:
          type: array
          items:
            type: string
            format: uuid
        state:
          $ref: '#/components/schemas/ReservationState'
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    CreateScreeningRequest:
      type: object
      properties:
        movieId:
          type: string
          format: uuid
        auditoriumId:
          type: string
          format: uuid
        pricingPlanId:
          type: string
          format: uuid
        refundPolicyId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
      required:
      - auditoriumId
      - endTime
      - movieId
      - pricingPlanId
      - refundPolicyId
      - startTime
    ScreeningPriceResponse:
      type: object
      properties:
        seatCategory:
          $ref: '#/components/schemas/SeatCategory'
        amount:
          type: number
        currency:
          type: string
    ScreeningResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        movieId:
          type: string
          format: uuid
        auditoriumId:
          type: string
          format: uuid
        pricingPlanId:
          type: string
          format: uuid
        refundPolicyId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/ScreeningStatus'
        prices:
          type: array
          items:
            $ref: '#/components/schemas/ScreeningPriceResponse'
        inventorySize:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
...
    CursorPageResponseScreeningSummary:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScreeningSummary'
        nextCursor:
          type: string
    ScreeningSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        movieId:
          type: string
          format: uuid
        auditoriumId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    PriceSummary:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/SeatCategory'
        amount:
          type: number
        currency:
          type: string
    ScreeningDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
        movieId:
          type: string
          format: uuid
        auditoriumId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        prices:
          type: array
          items:
            $ref: '#/components/schemas/PriceSummary'
        inventorySize:
          type: integer
          format: int32
    SeatAvailability:
      type: object
      properties:
        screeningSeatId:
          type: string
          format: uuid
        seatId:
          type: string
          format: uuid
        rowLabel:
          type: string
        seatNumber:
          type: integer
          format: int32
        category:
          $ref: '#/components/schemas/SeatCategory'
        state:
          $ref: '#/components/schemas/ScreeningSeatState'
...
    PageResponseScreeningResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScreeningResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    FieldViolation:
      type: object
      properties:
        field:
          type: string
          example: screeningSeatIds[1]
        code:
          type: string
          example: NotNull
        message:
          type: string
          example: The field is required.
      required:
      - code
      - field
      - message
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: /problems/seat-unavailable
        title:
          type: string
          example: Seat unavailable
        status:
          type: integer
          format: int32
          example: 409
        detail:
          type: string
          example: One or more requested seats are no longer available.
        instance:
          type: string
          format: uri
          example: /api/v1/seat-reservations
        code:
          type: string
          example: SEAT_UNAVAILABLE
        requestId:
          type: string
          format: uuid
          example: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
      required:
      - code
      - detail
      - instance
      - requestId
      - status
      - title
      - type
    SeatCategory:
      type: string
      enum:
      - REGULAR
      - PREMIUM
...
    ReservationState:
      type: string
      enum:
      - ACTIVE
      - RELEASED
      - EXPIRED
      - PAYMENT_IN_PROGRESS
      - CONVERTED
...
    ScreeningStatus:
      type: string
      enum:
      - ACTIVE
      - CANCELLED
    ScreeningSeatState:
      type: string
      enum:
      - AVAILABLE
      - RESERVED
      - PAYMENT_IN_PROGRESS
      - BOOKED
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Invalid request
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/invalid-request
            status: 400
            instance: /api/v1/seat-reservations
            code: INVALID_REQUEST
            detail: The request could not be processed.
    Unauthenticated:
      description: Valid HTTP Basic credentials are required.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Authentication required
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/authentication-required
            status: 401
            instance: /api/v1/seat-reservations
            code: AUTHENTICATION_REQUIRED
            detail: The request could not be processed.
    Forbidden:
      description: The authenticated account lacks the required role.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Access denied
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/access-denied
            status: 403
            instance: /api/v1/seat-reservations
            code: ACCESS_DENIED
            detail: The request could not be processed.
    NotFound:
      description: The resource does not exist or is concealed from this account.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Resource not found
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/resource-not-found
            status: 404
            instance: /api/v1/seat-reservations
            code: RESOURCE_NOT_FOUND
            detail: The request could not be processed.
    Conflict:
      description: The request conflicts with the current resource state.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Request conflict
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/request-conflict
            status: 409
            instance: /api/v1/seat-reservations
            code: REQUEST_CONFLICT
            detail: The request could not be processed.
    BusinessViolation:
      description: A business rule prevents the operation.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Business rule violation
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/business-rule-violation
            status: 422
            instance: /api/v1/seat-reservations
            code: BUSINESS_RULE_VIOLATION
            detail: The request could not be processed.
    InternalError:
      description: An unexpected failure occurred.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Internal server error
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/internal-error
            status: 500
            instance: /api/v1/seat-reservations
            code: INTERNAL_ERROR
            detail: The request could not be processed.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: Unique key used to safely retry a state-changing request
      required: true
      schema:
        type: string
        maxLength: 120
        minLength: 1
      example: demo-reservation-001
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        format: int32
        default: 0
    PageSize:
      name: size
      in: query
      description: Number of records from 1 to 100
      schema:
        type: integer
        format: int32
        default: 20
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the previous response
      schema:
        type: string
      example: eyJ2IjoxLCJrIjo...
    CursorLimit:
      name: limit
      in: query
      description: Number of records from 1 to 100
      schema:
        type: integer
        format: int32
        default: 20
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic